Optionally use environment PWD rather than Dir.pwd
When compiling Sass with grunt the extension is using the current working directory from where the grunttask was run rather than the directory the Sass lives in.  Normally, that's not a huge issue, but when trying to asychronously compile multiple projects using a single gruntfile then it's impossible to discern where the output is coming from without using different output files for each project.

Ex folder structure:
root\
|-- project_1\
|   |-- sass\
|       |-- something.sass
|   |-- config.rb
|-- project_2\
|   |-- sass\
|       |-- another_file.sass
|   |-- config.rb
|-- Gruntfile

    Running the Grunt task with the current setup will dump the output files into root\ when it would make sense to keep them in the folder where the sass is compiled from.  The proposed changes would output the SassExport files into the root\project_1\ and root\project_2\ folders respectively.